Introducing Vector storage utility (#18)
* feat(ai): add vector storage capabilities with PostgreSQL support

This commit adds vector storage functionality to the ai package with the following changes:

- Add vector storage interfaces and types for managing vector embeddings
- Implement PostgreSQL provider using pgvector extension
- Add vector validation utilities for data validation
- Change embedding type from float64 to float32 for compatibility
- Update dependencies to support vector storage features

The new vector storage system provides:
- Collection management (create, delete, list)
- Document operations (upsert, get, delete)
- Vector similarity search with configurable options
- Support for metadata and custom fields
- Validation for collections and documents
- Error handling with detailed error types

+++ ai/embedding.go
@@ -44,7 +44,7 @@ type EmbeddingObject struct {
 	// Object identifies the type of the response object
 	Object string `json:"object"`
 	// Embedding is the generated vector representation of the input text
-	Embedding []float64 `json:"embedding"`
+	Embedding []float32 `json:"embedding"`
 	// Index is the position of this embedding in the response array
 	Index int `json:"index"`
 }
